Ach... just another huge commit with everything

+++ src/Common/DbEntity/VisitedPlaceEntity.cs
@@ -7,8 +7,8 @@ namespace Gloobster.Common.DbEntity
 		public ObjectId PortalUser_id { get; set; }
 		public string CountryCode { get; set; }
 		public string City { get; set; }
-		public double PlaceLatitude { get; set; }
-		public double PlaceLongitude { get; set; }
+		public float PlaceLatitude { get; set; }
+		public float PlaceLongitude { get; set; }
 		public string SourceId { get; set; }
 		public string SourceType { get; set; }
 	}
