+++ src/DomainModelsCommon/DO/VisitedPlaceDO.cs
@@ -7,8 +7,8 @@ namespace Gloobster.DomainModelsCommon.DO
 		public string PortalUserId { get; set; }
 		public string CountryCode { get; set; }
 		public string City { get; set; }
-		public double PlaceLatitude { get; set; }
-		public double PlaceLongitude { get; set; }
+		public float PlaceLatitude { get; set; }
+		public float PlaceLongitude { get; set; }
 		public string SourceId { get; set; }
 		public SourceTypeDO SourceType { get; set; }
