+++ src/WebApiObjects/VisitedPlaceResponse.cs
@@ -10,8 +10,8 @@
 	{
 		public string CountryCode { get; set; }
 		public string City { get; set; }
-		public double PlaceLatitude { get; set; }
-		public double PlaceLongitude { get; set; }
+		public float PlaceLatitude { get; set; }
+		public float PlaceLongitude { get; set; }
 		public string SourceId { get; set; }
 		public string SourceType { get; set; }
 	}
